requirements.txt
Flask
scikit-learn
waitress
torch
torchvision
Pillow